// src/interfaces/dataInterface.ts
export interface Data {
  symbol: string;
  lastPrice: string;
  priceChangePercent: string;
  openPrice: string;
  closePrice: string;
  highPrice: string;
  lowPrice: string;
  volume: string;
  quoteVolume: string;
}

export interface CoinInfo {
  token: string;
  currency: string;
  name: string;
}